# internal/http/gen/model.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.7.1 DO NOT EDIT.
package gen

// Error defines model for Error.
type Error struct {
	Code    int32  `json:"code"`
	Message string `json:"message"`
}

// Pet defines model for Pet.
type Pet struct {
	Name string  `json:"name"`
	Tag  *string `json:"tag,omitempty"`
}

// PetResponse defines model for PetResponse.
type PetResponse struct {
	// Embedded struct due to allOf(#/components/schemas/Pet)
	Pet `yaml:",inline"`
	// Embedded fields due to inline allOf schema
	Id int64 `json:"id"`
}

// FindPetsParams defines parameters for FindPets.
type FindPetsParams struct {

	// tags to filter by
	Tags *[]string `json:"tags,omitempty"`

	// maximum number of results to return
	Limit *int32 `json:"limit,omitempty"`

	// maximum number of results to return
	Order *string `json:"order,omitempty"`
}

// AddPetJSONBody defines parameters for AddPet.
type AddPetJSONBody Pet

// AddPetJSONRequestBody defines body for AddPet for application/json ContentType.
type AddPetJSONRequestBody AddPetJSONBody